+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Cache authentication info ([#15](https://github.com/scm-manager/scm-ldap-plugin/pull/15))
+
 ## 2.1.0 - 2021-02-15
 ### Added
 - Nested group support for non-AD servers ([#13](https://github.com/scm-manager/scm-ldap-plugin/pull/13))